# swagger.yaml
swagger: '2.0'
basePath: /api/v0.1.1
info:
  title: WP3-backed Data Service API
  description: An API which supports querying available LEXIS data sets
  version: 0.1.1
  contact:
    email: ruben@lrz.de
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
definitions:
...
  ErrorResponse:
    type: object
    required:
      - errorString
    properties:
      errorString:
        type: string
...
securityDefinitions:
  Keycloak:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https:138.246.232.245:6443/auth'
    tokenUrl: 'https:138.246.232.245:6443/auth'
security:
  - Keycloak: []
paths:
...
  /staging/stage/{request_id}:
    get:
      security: []
      tags:
        - staging
        - dataSetManagement
      produces:
        - application/json
      summary: Check the status of a transfer
      operationId: CheckStageStatus
      description: Check the status of a transfer
      parameters:
        - in: path
          name: request_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: This means that the status has been returned to the user in the response body.
          schema:
            type: object
            required: 
              - status
            properties:
              status: 
                type: string
              target_path:
                type: string
                description: Single path for staging endpoints
        '400':
          description: This means that the request ID given by the user is incorrect.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: This means that the ID doesn't exist and thus a status can't be returned.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the the request ID is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /staging/delete/{request_id}:
    get:
      security: []
      tags:
        - staging
        - dataSetManagement
      produces:
        - application/json
      summary: Check the status of a delete request
      operationId: CheckDeletionStatus
      description: Check the status of a delete request
      parameters:
        - in: path
          name: request_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: This means that the status has been returned to the user in the response body.
          schema:
            type: object
            required: 
              - status
            properties:
              status: 
                type: string
        '400':
          description: This means that the request ID given by the user is incorrect.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: This means that the ID doesn't exist and thus a status can't be returned.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the the request ID is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
...